Make math headers explicit and enable unit tests on Windows
The collision, camera, and particle code called bare C math functions
(fabsf, sqrtf, cosf, sinf, expf, floorf) without including a math header,
relying on those names leaking into the global namespace transitively.
That happens to hold with libstdc++ on Linux but is not guaranteed under
MSVC, where the global-namespace C math names may be absent, so the gem
could fail to compile on Windows (O3DE's primary platform).

Include <cmath> in each file and call the std:: forms. These are the same
libm functions, so numerical behavior is byte-identical to the previous
calls (all 136 unit tests, including the collision and camera math,
still pass); the change is purely making the dependency explicit and
portable across compilers.

Also enable PAL_TRAIT_DIORAMA_TEST_SUPPORTED on Windows. The unit tests
are pure platform-agnostic C++ with no Linux dependency, so a Windows
host build now gets the same test coverage instead of skipping the
Tests target.

Note: the gem still needs a real MSVC build to confirm end to end; this
removes the one cross-compiler hazard found in a source audit.

+++ Code/Source/Clients/Camera2D.h
@@ -10,6 +10,8 @@
 #include <AzCore/Math/Vector2.h>
 #include <AzCore/base.h>
 
+#include <cmath>
+
 // Pure, header-only math for the 2D camera component (Docs/design/2d-camera.md).
 // These are the frame-rate-independent building blocks of camera follow: smoothing,
 // deadzone, bounds, lookahead, pixel snapping, and trauma-based screen shake. They
@@ -39,7 +41,7 @@ namespace Diorama::Camera2D
         {
             return target;
         }
-        const float factor = 1.0f - expf(-deltaTime / smoothTime);
+        const float factor = 1.0f - std::exp(-deltaTime / smoothTime);
         return current + (target - current) * factor;
     }
 
@@ -109,8 +111,8 @@ namespace Diorama::Camera2D
         {
             return pos;
         }
-        const float x = floorf(pos.GetX() * pixelsPerUnit + 0.5f) / pixelsPerUnit;
-        const float y = floorf(pos.GetY() * pixelsPerUnit + 0.5f) / pixelsPerUnit;
+        const float x = std::floor(pos.GetX() * pixelsPerUnit + 0.5f) / pixelsPerUnit;
+        const float y = std::floor(pos.GetY() * pixelsPerUnit + 0.5f) / pixelsPerUnit;
         return AZ::Vector2(x, y);
     }
 
@@ -143,6 +145,6 @@ namespace Diorama::Camera2D
     //! supplies the angle from smooth noise so this stays deterministic and testable.
     inline AZ::Vector2 ShakeOffset(float magnitude, float angle)
     {
-        return AZ::Vector2(cosf(angle) * magnitude, sinf(angle) * magnitude);
+        return AZ::Vector2(std::cos(angle) * magnitude, std::sin(angle) * magnitude);
     }
 } // namespace Diorama::Camera2D
